fix:fix glm2

+++ examples/chatglm2_6b/run_parallel.sh
@@ -11,19 +11,24 @@ RUN_OPTION_LIST="--run --performance --webdemo --zhipu --profiling"
 RUN_OPTION="--run"
 SCRIPT_PATH=$SCRIPT_DIR/modeling_chatglm.py
 OUTPUT_DIR="tensor_parallel"
+export HCCL_BUFFSIZE=110
+export ACLTRANSFORMER_PLAN_EXECUTE_ASYNC=1
+export TASK_QUEUE_ENABLE=1
+export HCCL_OP_BASE_FFTS_MODE_ENABLE=1
+export ACLTRANSFORMER_CONVERT_NCHW_TO_ND=1
 
 function fn_prepare()
 {
     echo "$RUN_OPTION $SCRIPT_PATH"
 
     if [ ! -f "$MODEL_TARGET_DIR/pytorch_model-00001-of-00007.bin" ];then
-        ln -s $ACLTRANSFORMER_TESTDATA/weights/chatglm2_6b/pytorch_model-00001-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00001-of-00007.bin
-        ln -s $ACLTRANSFORMER_TESTDATA/weights/chatglm2_6b/pytorch_model-00002-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00002-of-00007.bin
-        ln -s $ACLTRANSFORMER_TESTDATA/weights/chatglm2_6b/pytorch_model-00003-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00003-of-00007.bin
-        ln -s $ACLTRANSFORMER_TESTDATA/weights/chatglm2_6b/pytorch_model-00004-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00004-of-00007.bin
-        ln -s $ACLTRANSFORMER_TESTDATA/weights/chatglm2_6b/pytorch_model-00005-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00005-of-00007.bin
-        ln -s $ACLTRANSFORMER_TESTDATA/weights/chatglm2_6b/pytorch_model-00006-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00006-of-00007.bin
-        ln -s $ACLTRANSFORMER_TESTDATA/weights/chatglm2_6b/pytorch_model-00007-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00007-of-00007.bin
+        ln -s $ACLTRANSFORMER_TESTDATA/pytorch_model-00001-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00001-of-00007.bin
+        ln -s $ACLTRANSFORMER_TESTDATA/pytorch_model-00002-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00002-of-00007.bin
+        ln -s $ACLTRANSFORMER_TESTDATA/pytorch_model-00003-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00003-of-00007.bin
+        ln -s $ACLTRANSFORMER_TESTDATA/pytorch_model-00004-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00004-of-00007.bin
+        ln -s $ACLTRANSFORMER_TESTDATA/pytorch_model-00005-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00005-of-00007.bin
+        ln -s $ACLTRANSFORMER_TESTDATA/pytorch_model-00006-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00006-of-00007.bin
+        ln -s $ACLTRANSFORMER_TESTDATA/pytorch_model-00007-of-00007.bin $MODEL_TARGET_DIR/pytorch_model-00007-of-00007.bin
     fi
 
 
@@ -55,7 +60,9 @@ function fn_prepare_parallel()
 
 function fn_clean()
 {
-    rm $TRANSFORMER_PACKAGE_PATH/configuration_chatglm.py
+    if [ -f $TRANSFORMER_PACKAGE_PATH/configuration_chatglm.py ];then
+        rm $TRANSFORMER_PACKAGE_PATH/configuration_chatglm.py
+    fi
 }
 
 function fn_main()
@@ -69,42 +76,68 @@ function fn_main()
         fi
     fi
     
+    # get option
     if [[ ! -z "$1" ]];then
         RUN_OPTION=$1
     fi
     cd $SCRIPT_DIR
 
-    fn_prepare
-    fn_prepare_parallel
+    # cut weights first
+    if [ ! -d "$OUTPUT_DIR" ] && [ "$RUN_OPTION" != "--cut" ];then
+        echo "The weight has not cut, please cut weight first:"
+        echo "run_parallel.sh [model script path] [--cut]"
+        exit
+    fi
+
+    if [ -d "$OUTPUT_DIR" ];then
+        if [ "$RUN_OPTION" == "--cut" ]; then
+            echo "The cut weights has exist, please remove weights or choose other options:"
+            echo "run_parallel.sh [model script path] [--performance|--precision <output_name>|--zhipu]"
+            exit
+        else
+            fn_prepare_parallel
+        fi
+    fi
 
     case "${RUN_OPTION}" in
         "--cut")
+            fn_prepare
             python3 $SCRIPT_DIR/cut_model_util.py --output_path $OUTPUT_DIR
             ;;
         "--performance")
-            torchrun --nproc_per_node 2 --master_port 20000 $SCRIPT_DIR/main_performance.py --parallel
+            torchrun --nproc_per_node 2 --master_port 20000 $SCRIPT_DIR/main_performance.py --parallel --model_path $OUTPUT_DIR
             ;;
+
+        "--multi-batch")
+            torchrun --nproc_per_node 2 --master_port 20000 $SCRIPT_DIR/main_batch_performance.py --parallel --model_path $OUTPUT_DIR
+            ;;
+
+        # TODO
         # "--webdemo")
         #     unset https_proxy
         #     unset http_proxy
         #     python3 $SCRIPT_DIR/main_web.py
         #     ;;
+
         "--zhipu")
-            torchrun --nproc_per_node 2 --master_port 20000 $SCRIPT_DIR/zhipu_test.py --parallel
+            torchrun --nproc_per_node 2 --master_port 20000 $SCRIPT_DIR/zhipu_test.py --parallel --model_path $OUTPUT_DIR
             ;;
-        # "--profiling")
-            
+        
+        # TODO
+        # "--profiling")  
         #     ;;
-        # "--precision")
-        #     python3 $SCRIPT_DIR/main_precision.py $2
-        #     ;;
+
+        "--precision")
+            python3 $SCRIPT_DIR/main_precision.py $2
+            ;;
+        
         "--help")
             echo "export ACLTRANSFORMER_TESTDATA=<path>"
-            echo "run.sh [model script path] [--performance|--precision <output_name>|--zhipu]"
+            echo "run_parallel.sh [model script path] [--cut|--performance|--precision <output_name>|--zhipu]"
             ;;
         *)
             echo "unknown build type:${RUN_OPTION}"
-            echo "run.sh [model script path] [--performance|--precision <output_name>|--zhipu]"
+            echo "run_parallel.sh [model script path] [--cut|performance|--precision <output_name>|--zhipu]"
             ;;
     esac
